backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
supabase==2.0.3
sqlalchemy==2.0.23
pydantic==2.5.2
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2  # Para ML si lo necesitas
